moved everything into MainLayout for now
