--- RoyAppMaui/Components/Pages/SleepTable.razor
@@ -1,80 +1,0 @@
-﻿@page "/"
-
-<MudDataGrid T="Sleep"
-             Bordered="true"
-             Dense="true"
-             Elevation="25"
-             FixedHeader="true"
-             FixedFooter="true"
-             Height="calc(100vh - 220px)"
-             Items="_items"
-             Loading="@_isLoading"
-             LoadingProgressColor="@Color.Primary"
-             ReadOnly="false"
-             RowsPerPage="Settings.RowsPerPage"
-             RowsPerPageChanged="OnRowsPerPageChanged"
-             CommittedItemChanges="OnCommittedItemChanges">
-    <Columns>
-        <PropertyColumn Property="x => x.Id" />
-        <PropertyColumn Property="x => x.BedtimeDisplay" Title="Bedtime" />
-        <PropertyColumn Property="x => x.BedtimeRec" Editable="false" Title="(As Decimal)" />
-        <PropertyColumn Property="x => x.WaketimeDisplay" Title="Waketime" />
-        <PropertyColumn Property="x => x.WaketimeRec" Editable="false" Title="(As Decimal)" />
-        <PropertyColumn Property="x => x.Duration" Editable="false" />
-        <TemplateColumn Filterable="false" Sortable="false">
-            <CellTemplate>
-                <MudIconButton Size="@Size.Small"
-                               Icon="@Icons.Material.Filled.Edit"
-                               OnClick="@(() => OnModifyItem(context.Item))" />
-                <MudIconButton Size="@Size.Small"
-                               Icon="@Icons.Material.Filled.Delete"
-                               Color="Color.Error"
-                               OnClick="@(() => RemoveItemAsync(context.Item))" />
-            </CellTemplate>
-        </TemplateColumn>
-    </Columns>
-    <PagerContent>
-        <MudDataGridPager T="Sleep" />
-    </PagerContent>
-</MudDataGrid>
-
-<MudDivider />
-
-<MudPaper Width="100%" Height="102px" Square="true">
-    <MudStack class="ml-4" Row="true" Style="width: 430px">
-        <AverageField Label="Bedtime Avg" Value="@BedtimeAvg" />
-        <AverageField Label="Waketime Avg" Value="@WaketimeAvg" />
-        <MudButton Class="mt-4" Variant="Variant.Filled" StartIcon="@Icons.Material.Filled.PlaylistRemove" Color="Color.Primary" OnClick="@ClearTable">Clear Data</MudButton>
-    </MudStack>
-    <MudSpacer />
-    <MudItem Class="fab-fixed-position d-flex justify-end flex-grow-1 gap-4 mr-5 mb-3">
-        <MudFab Color="Color.Success"
-                OnClick="ShowAddNewItemOverlay"
-                StartIcon="@Icons.Material.Filled.Add" />
-    </MudItem>
-</MudPaper>
-
-<MudOverlay @bind-Visible="_showModifyItemDialog" DarkBackground>
-    <MudPaper Class="pa-4" Style="width:400px;">
-        <MudForm @ref="itemForm" @bind-IsValid="@isValidItem">
-            <MudText Class="mb-5" Typo="Typo.h6">Modify Bedtime</MudText>
-
-            <MudTextField @bind-Value="@_selectedItem.Id" Label="Id" Variant="Variant.Outlined" Required RequiredError="An Id must be supplied!" />
-
-            <MudStack Row="true">
-                <MudTimePicker AmPm="true" Label="Bedtime" Editable="true" Variant="Variant.Outlined" Time="_selectedItem.Bedtime" TimeChanged="args => HandleTimeChange(TimePickers.Bedtime, args)" />
-                <MudField Class="mud-field-decimal-display" Label="Decimal" Variant="Variant.Outlined">@_selectedItem.BedtimeRec</MudField>
-            </MudStack>
-
-            <MudStack Row="true">
-                <MudTimePicker AmPm="true" Label="Waketime" Editable="true" Variant="Variant.Outlined" Time="_selectedItem.Waketime" TimeChanged="args => HandleTimeChange(TimePickers.Waketime, args)" />
-                <MudField Class="mud-field-decimal-display" Label="Decimal" Variant="Variant.Outlined">@_selectedItem.WaketimeRec</MudField>
-            </MudStack>
-
-            <MudStack Class="mt-5" Row="true" Justify="Justify.FlexEnd">
-                <MudButton OnClick="@OnSaveBedtimeItem" Variant="Variant.Filled" Color="Color.Primary" StartIcon="@Icons.Material.Filled.Save">Save</MudButton>
-                <MudButton OnClick="@OnCloseOverlay" Variant="Variant.Filled">Cancel</MudButton>
-            </MudStack>
-        </MudForm>
-    </MudPaper>
-</MudOverlay>
