+++ RoyAppMaui/wwwroot/css/app.css
@@ -55,7 +55,7 @@
 }
 
 .mud-field-decimal-display {
-    width: 60px !important;
+    width: 78px !important;
     text-align: center !important;
 }
